Addition of Wifi Direct (simulated) functionality.
- Renamed SelectUserActivity to SettingsActivity and added a toggle
button to enabled/disable Wifi Direct.

- Deleted IMenuActivity and IGameActivity. Not needed.

- Added WDSim related code to MultiplayerMenuActivity and GameActivity.

- Package structure and naming was modified to make a bit more sense.
Package network holds everything that deals with ... networking.

- ClientService was renamed to NetworkService.

- ClientAsyncTask was modified to work with an additional type of
sockets and also to receive a list of addresses of the peers in the P2P
group.

- Added server code (previously on dropbox) so any player can now also
be the server. Changed only the socket type it uses, to work with WDSim.

- Added WDSim lib to libs (necessary to build project).

- Code refactoring all over the project.

NOTE: As it is now, it should be possible to play a full game. Every
functionality that worked with a central server should still work. It
assumes that the server is ALWAYS at 192.168.0.1. If you start the
server at another address, it will not work (work in progress).